requirements.txt
argcomplete==1.12.3
asgiref==3.4.1
boto3==1.19.12
botocore==1.22.12
certifi==2021.10.8
cffi==1.15.0
cfn-flip==1.3.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
Deprecated==1.2.13
Django==3.2.9
djangorestframework==3.12.4
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==3.3
install==1.3.4
jmespath==0.10.0
kappa==0.6.0
pep517==0.12.0
pip-tools==6.4.0
placebo==0.10.0
psycopg2-binary==2.9.1
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.4.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.2
tqdm==4.62.3
troposphere==3.1.0
urllib3==1.26.7
Werkzeug==0.16.1
wrapt==1.13.3
wsgi-request-logger==0.4.6
zappa==0.54.0
